Using test properties file for testing the LoadAttachmentTask.
We have added a filepath dependency injection in the short term.  In the long term we will be
adding full dependency injection of all supporting utilities, so the PropertyFileUtil can be
properly mocked out during testing (which will make this commit no longer use the actual files).
However, for right now that is outside the scope of this ticket, and the use of a test file is our
best bet for this current feature.