Fixed the bug where one or more of the first rows failed due to an exception
The error displayed is: ConcurrentModificationException, which arose because of the
way we are using a static factory pattern to grab a single instance of an EntityAssociations
object and then calling the allAssociations() method on that instance from multiple
threads concurrently. The creation of the array of values is what fails.